2/12/25 Minor Adjustments Reverted unnecessary changes related to switching between robot and field centric Modified math for the dpad in twoPlayerTeleop.java Minor improvements / changes to classes in tuning Properly grouped certain files when viewing them from driver hub Removed redundant FieldCentric class from laboratory Created new distanceSensor_lab class for potential work with distance sensors General improvements to documentation Improved readability of some math
Stuff that still has to be done:
TESTING! testing teleop, other changes
linear_otos.java and angular_otos.java seem to be incomplete, with the only functions of them outputting position telemetry. They still need code that can be used for tuning.
For drive.java's gotoPos, I am uncertain of a lot of the math and other stuff that is used in it, but some parts, particularly related to Vector2D.java, seem somewhat convoluted. The math for setting the power is probably wrong, and it does not seem to have a way to keep bounds, although I do not know if the math in the pidHeading and pdfTranslate fulfills that need. It may be better to adapt the code to use the RobotCentricDrive function instead.
We do not seem to have an auto that is even close to usable, and functions that would be used for it like gotoPos may be incomplete or nonfunctional. Significant work is needed, and this is probably a priority over most of everything else.